Minor cleanup and change some functions/structs to internal
